+++ Models/ExtraFees.cs
@@ -2,9 +2,10 @@
 {
     public class ExtraFees
     {
-        // may need to use an array here...
+        //public ExtraFeesArray ExtraFeesArray { get; set; } = default!;
+
         public string ExtraFeeId { get; set; } = default!;
         public int Quantity { get; set; } = default!;
-        public decimal UnitPrice { get; set; } = default!;
+        public double UnitPrice { get; set; } = default!;
     }
 }
