+++ Models/PickupLocation.cs
@@ -2,8 +2,8 @@
 {
     public class PickupLocation
     {
-        public decimal Latitude { get; set; } = default!;
+        public double Latitude { get; set; } = default!;
 
-        public decimal Longitude { get; set; } = default!;
+        public double Longitude { get; set; } = default!;
     }
 }
